+++ src/main/java/paul/fallen/module/modules/combat/AutoTNT.java
@@ -112,7 +112,7 @@ public final class AutoTNT extends Module {
                 PlayerControllerUtils.rightClickBlock(new Vector3d(0.5, 0.5, 0.5), Direction.DOWN, tntPos);
                 mc.player.swingArm(Hand.MAIN_HAND);
 
-                int[] rot = RotationUtils.getYawAndPitch(new Vector3d(tntPos.getX() + 0.5, tntPos.getY() + 0.5, tntPos.getZ() + 0.5));
+                float[] rot = RotationUtils.getYawAndPitch(new Vector3d(tntPos.getX() + 0.5, tntPos.getY() + 0.5, tntPos.getZ() + 0.5));
                 if (mode.sval == "packet") {
                     mc.player.connection.sendPacket(new CPlayerPacket.RotationPacket(rot[0], rot[1], mc.player.isOnGround()));
                 } else {
@@ -128,7 +128,7 @@ public final class AutoTNT extends Module {
                 PlayerControllerUtils.rightClickBlock(new Vector3d(0.5, 0, 0.5), Direction.DOWN, tntPos);
                 mc.player.swingArm(Hand.MAIN_HAND);
 
-                int[] rot = RotationUtils.getYawAndPitch(new Vector3d(tntPos.getX() + 0.5, tntPos.getY(), tntPos.getZ() + 0.5));
+                float[] rot = RotationUtils.getYawAndPitch(new Vector3d(tntPos.getX() + 0.5, tntPos.getY(), tntPos.getZ() + 0.5));
                 if (mode.sval == "packet") {
                     mc.player.connection.sendPacket(new CPlayerPacket.RotationPacket(rot[0], rot[1], mc.player.isOnGround()));
                 } else {
